mcp_study/pyproject.toml
[project]
name = "mcp_study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.10.6",
    "httpx>=0.28.1",
    "mcp[cli]>=1.11.0",
    "openai>=1.97.0",
    "python-dotenv>=1.1.1",
]
